fix(http-elixir): Fix stateful cooldown
Elixir goes through a number of actions to start the network service. It
requires more that 1000ms for proper stateful start. Increase cooldown
to `3000ms`.

Note that after starting an Elixir instance, several seconds are
required for it to reach the final state. During that time, requests
will fail. After that, queries will work as expected, in a stateful
setup.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>